+++ core/src/main/java/org/musetest/core/step/CallFunction.java
@@ -3,6 +3,7 @@ package org.musetest.core.step;
 import org.musetest.core.*;
 import org.musetest.core.context.*;
 import org.musetest.core.events.*;
+import org.musetest.core.resource.*;
 import org.musetest.core.step.descriptor.*;
 import org.musetest.core.steptest.*;
 import org.musetest.core.values.*;
@@ -34,7 +35,7 @@ import java.util.*;
 public class CallFunction extends CallMacroStep
     {
     @SuppressWarnings("unused") // called via reflection
-    public CallFunction(StepConfiguration config, MuseProject project) throws StepConfigurationError
+    public CallFunction(StepConfiguration config, MuseProject project) throws MuseInstantiationException
         {
         super(config, project);
         _config = config;